Move counter tests to a dedicated file
